Semi-automatic code cleanup * Unused imports * Redundant semicolons * Redundant boxing/unboxing * If statement simplification * Foreach instead of for * Javadoc @link errors * Redundant escapes in regex * Optional with null cleanup * Never-failing assertions * !.anyMatch to .noneMatch * catch immediate rethrow * .indexOf to .contains

+++ community/graph-algo/src/test/java/org/neo4j/graphalgo/impl/util/TestPriorityMap.java
@@ -92,7 +92,7 @@ public class TestPriorityMap
         Entry<Integer, Double> top = map.pop();
         assertNull( map.peek() );
         assertEquals( 1, top.getEntity().intValue() );
-        assertEquals( 1.5d, top.getPriority().doubleValue(), 0d );
+        assertEquals( 1.5d, top.getPriority(), 0d );
     }
 
     @Test
